fix(db): fail closed on column migration errors
